Safari-611.1.10/LayoutTests/http/tests/security/mixedContent/secure-redirect-to-secure-redirect-to-basic-auth-insecure-image-allowCrossOriginSubresourcesToAskForCredentials.https-expected.txt

CONSOLE MESSAGE: The page at https://127.0.0.1:8443/security/mixedContent/secure-redirect-to-secure-redirect-to-basic-auth-insecure-image-allowCrossOriginSubresourcesToAskForCredentials.https.html was allowed to display insecure content from http://localhost:8080/security/mixedContent/resources/subresource/protected-image.php.

https://127.0.0.1:8443/resources/redirect.php?url=https%3A//localhost%3A8443/resources/redirect.php%3Furl%3Dhttp%3A//localhost%3A8080/security/mixedContent/resources/subresource/protected-image.php - didReceiveAuthenticationChallenge - Responding with testUser:testPassword
This test loads a secure image that redirects to an secure image that redirects to an insecure image guarded by basic authentication. The image should load.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS did load image.
PASS successfullyParsed is true

TEST COMPLETE
